# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WealthApp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /budget:
    get:
      operationId: listBudgetEntries
      tags: [budget]
      summary: List budget entries for current user
      responses:
        "200":
          description: Budget entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetEntry"
    post:
      operationId: createBudgetEntry
      tags: [budget]
      summary: Create a budget entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetEntry"
  /budget/{id}:
    put:
      operationId: updateBudgetEntry
      tags: [budget]
      summary: Update a budget entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetEntry"
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals for current user
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialGoal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialGoal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialGoal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /kyc/documents:
    get:
      operationId: listKycDocuments
      tags: [kyc]
      summary: List KYC documents
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KycDocument"
    post:
      operationId: uploadKycDocument
      tags: [kyc]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadKycDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycDocument"
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    put:
      operationId: updateLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}/admin:
    put:
      operationId: updateBlogPost
      tags: [blog]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
...
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: adminUpdateUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /advisor/clients/{id}/goals:
    get:
      operationId: advisorGetClientGoals
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialGoal"
...
  /advisor/clients/{id}/kyc:
    get:
      operationId: advisorGetClientKyc
      tags: [kyc]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: KYC docs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KycDocument"
  /advisor/clients/{id}/kyc/{docId}/verify:
    post:
      operationId: advisorVerifyKycDocument
      tags: [kyc]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: docId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyKycDocumentBody"
      responses:
        "200":
          description: Updated doc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycDocument"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        preferredCurrency:
          type: string
        countryCode:
          type: ["string", "null"]
        isExpat:
          type: boolean
        onboardingComplete:
          type: boolean
        riskProfile:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, email, role, preferredCurrency, isExpat, onboardingComplete, createdAt, updatedAt]
...
    BudgetEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        periodMonth:
          type: string
        currency:
          type: string
        income:
          type: ["string", "null"]
        needsActual:
          type: ["string", "null"]
        wantsActual:
          type: ["string", "null"]
        savingsActual:
          type: ["string", "null"]
        savingsRatePercent:
          type: ["string", "null"]
        housing:
          type: ["string", "null"]
        food:
          type: ["string", "null"]
        transport:
          type: ["string", "null"]
        utilities:
          type: ["string", "null"]
        entertainment:
          type: ["string", "null"]
        other:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, periodMonth, currency, createdAt, updatedAt]
    CreateBudgetEntryBody:
      type: object
      properties:
        periodMonth:
          type: string
        currency:
          type: string
        income:
          type: string
        housing:
          type: string
        food:
          type: string
        transport:
          type: string
        utilities:
          type: string
        entertainment:
          type: string
        other:
          type: string
        notes:
          type: string
      required: [periodMonth]
    FinancialGoal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        goalType:
          type: string
        targetAmount:
          type: ["string", "null"]
        currentAmount:
          type: string
        monthlyContribution:
          type: string
        targetDate:
          type: ["string", "null"]
        currency:
          type: string
        priority:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        isAdvisorManaged:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, title, goalType, currentAmount, monthlyContribution, currency, priority, status, isAdvisorManaged, createdAt, updatedAt]
    CreateGoalBody:
      type: object
      properties:
        title:
          type: string
        goalType:
          type: string
        targetAmount:
          type: string
        currentAmount:
          type: string
        monthlyContribution:
          type: string
        targetDate:
          type: string
        currency:
          type: string
        priority:
          type: string
        status:
          type: string
        notes:
          type: string
      required: [title, goalType]
...
    KycDocument:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        documentType:
          type: string
        fileUrl:
          type: string
        fileName:
          type: ["string", "null"]
        status:
          type: string
        rejectionReason:
          type: ["string", "null"]
        uploadedAt:
          type: string
        verifiedBy:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
      required: [id, userId, documentType, fileUrl, status, uploadedAt]
    UploadKycDocumentBody:
      type: object
      properties:
        documentType:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
      required: [documentType, fileUrl]
    VerifyKycDocumentBody:
      type: object
      properties:
        status:
          type: string
        rejectionReason:
          type: string
      required: [status]
...
    Lead:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        quizResult:
          type: ["string", "null"]
        healthScore:
          type: ["integer", "null"]
        status:
          type: string
        assignedAdvisorId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, email, status, createdAt, updatedAt]
    CreateLeadBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        source:
          type: string
        quizResult:
          type: string
        healthScore:
          type: integer
        status:
          type: string
        notes:
          type: string
      required: [email]
    BlogPost:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: ["string", "null"]
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        contentMarkdown:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        publishedAt:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, status, createdAt, updatedAt]
    CreateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        contentMarkdown:
          type: string
        category:
          type: string
        status:
          type: string
        coverImageUrl:
          type: string
      required: [title, slug]
...
    AdminUpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
        advisorId:
          type: string
        preferredCurrency:
          type: string